2019.4.28
	重写网络相关配置存储
	解决V2.3版键盘环绕灯指示异常的bug
	加入删除任务函数

+++ SOFTWARE/CScript/cScript.c
@@ -715,7 +715,7 @@ u32 findFunAddr (char *fun_name)
 	{
 		if (samestr((u8*)FUN_LIST[i].fun_name,(u8*)fun_name))
 		{
-			return FUN_LIST[i].fun_addr;
+			return (u32)FUN_LIST[i].fun_addr;
 		}
 	}
 	return 0;
